refactor memo module
